+++ src/abstractdataframe/reshape.jl
@@ -200,15 +200,15 @@ end
     unstack(df::AbstractDataFrame, rowkeys, colkey, value;
             renamecols::Function=identity, allowmissing::Bool=false,
             allowduplicates::Bool=false, valuestransform=nothing,
-            fill=missing, multithreaded::Bool=true)
+            fill=missing, threads::Bool=true)
     unstack(df::AbstractDataFrame, colkey, value;
             renamecols::Function=identity, allowmissing::Bool=false,
             allowduplicates::Bool=false, valuestransform=nothing,
-            fill=missing, multithreaded::Bool=true)
+            fill=missing, threads::Bool=true)
     unstack(df::AbstractDataFrame;
             renamecols::Function=identity, allowmissing::Bool=false,
             allowduplicates::Bool=false, valuestransform=nothing,
-            fill=missing, multithreaded::Bool=true)
+            fill=missing, threads::Bool=true)
 
 Unstack data frame `df`, i.e. convert it from long to wide format.
 
@@ -244,7 +244,7 @@ Row and column keys will be ordered in the order of their first appearance.
   default is `missing`. If the `value` column is a `CategoricalVector` and
   `fill` is not `missing` then in order to keep unstacked value columns also
   `CategoricalVector` the `fill` must be passed as `CategoricalValue`
-  - `multithreaded`: whether `valuestransform` may be run in separate tasks which
+  - `threads`: whether `valuestransform` may be run in separate tasks which
     can execute in parallel (possibly being applied to multiple groups at the same time).
     Whether or not tasks are actually spawned and their number are determined automatically.
     Set to `false` if `valuestransform` requires serial execution or is not thread-safe.
@@ -401,7 +401,7 @@ function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
                  values::ColumnIndex; renamecols::Function=identity,
                  allowmissing::Bool=false, allowduplicates::Bool=false,
                  valuestransform=nothing, fill=missing,
-                 multithreaded::Bool=true)
+                 threads::Bool=true)
     # first make sure that rowkeys are unique and
     # normalize all selectors as a strings
     # if some of the selectors are wrong we will get an early error here
@@ -434,7 +434,7 @@ function unstack(df::AbstractDataFrame, rowkeys, colkey::ColumnIndex,
             agg_fun = Ref∘valuestransform
         end
         df_op = combine(gdf, values => agg_fun => values_out,
-                        multithreaded=multithreaded)
+                        threads=threads)
 
         group_rows = find_group_row(gdf)
         if !issorted(group_rows)
@@ -459,22 +459,22 @@ function unstack(df::AbstractDataFrame, colkey::ColumnIndex, values::ColumnIndex
                  renamecols::Function=identity,
                  allowmissing::Bool=false, allowduplicates::Bool=false,
                  valuestransform=nothing, fill=missing,
-                 multithreaded::Bool=true)
+                 threads::Bool=true)
     colkey_int = index(df)[colkey]
     value_int = index(df)[values]
     return unstack(df, Not(colkey_int, value_int), colkey_int, value_int,
             renamecols=renamecols, allowmissing=allowmissing,
             allowduplicates=allowduplicates, valuestransform=valuestransform,
-            fill=fill, multithreaded=multithreaded)
+            fill=fill, threads=threads)
 end
 
 unstack(df::AbstractDataFrame; renamecols::Function=identity,
         allowmissing::Bool=false, allowduplicates::Bool=false,
         valuestransform=nothing, fill=missing,
-        multithreaded::Bool=true) =
+        threads::Bool=true) =
     unstack(df, :variable, :value, renamecols=renamecols, allowmissing=allowmissing,
             allowduplicates=allowduplicates, valuestransform=valuestransform,
-            fill=fill, multithreaded=multithreaded)
+            fill=fill, threads=threads)
 
 # we take into account the fact that idx, starts and ends are computed lazily
 # so we rather directly reference the gdf.groups