+++ src/groupeddataframe/splitapplycombine.jl
@@ -26,7 +26,7 @@ function _combine_prepare(gd::GroupedDataFrame,
                           (cs,)::Ref{Any};
                           keepkeys::Bool, ungroup::Bool, copycols::Bool,
                           keeprows::Bool, renamecols::Bool,
-                          multithreaded::Bool)
+                          threads::Bool)
     for cei in cs
         if !(cei isa AbstractMatrix && isempty(cei))
             @assert cei isa Union{Pair, Base.Callable, ColumnIndex,
@@ -49,14 +49,14 @@ function _combine_prepare(gd::GroupedDataFrame,
         end
     end
     return _combine_prepare_norm(gd, cs_vec, keepkeys, ungroup, copycols,
-                                 keeprows, renamecols, multithreaded)
+                                 keeprows, renamecols, threads)
 end
 
 function _combine_prepare_norm(gd::GroupedDataFrame,
                                cs_vec::Vector{Any},
                                keepkeys::Bool, ungroup::Bool, copycols::Bool,
                                keeprows::Bool, renamecols::Bool,
-                               multithreaded::Bool)
+                               threads::Bool)
     if any(x -> x isa Pair && first(x) isa Tuple, cs_vec)
         x = cs_vec[findfirst(x -> first(x) isa Tuple, cs_vec)]
         # an explicit error is thrown as this was allowed in the past
@@ -84,7 +84,7 @@ function _combine_prepare_norm(gd::GroupedDataFrame,
     # if earlier column with a column with the same name was created
 
     idx, valscat = _combine(gd, cs_norm, optional_transform, copycols, keeprows,
-                            renamecols, multithreaded)
+                            renamecols, threads)
 
     !keepkeys && ungroup && return valscat
 
@@ -361,12 +361,12 @@ function _combine_process_callable(wcs_i::Ref{Any},
                                    seen_cols::Dict{Symbol, Tuple{Bool, Int}},
                                    trans_res::Vector{TransformationResult},
                                    idx_agg::Ref{Vector{Int}},
-                                   multithreaded::Bool)
+                                   threads::Bool)
     cs_i = only(wcs_i)
     @assert cs_i isa Base.Callable
     firstres = length(gd) > 0 ? cs_i(gd[1]) : cs_i(similar(parentdf, 0))
     idx, outcols, nms = _combine_multicol(Ref{Any}(firstres), wcs_i, gd,
-                                          Ref{Any}(nothing), multithreaded)
+                                          Ref{Any}(nothing), threads)
 
     if !(firstres isa Union{AbstractVecOrMat, AbstractDataFrame,
                             NamedTuple{<:Any, <:Tuple{Vararg{AbstractVector}}}})
@@ -416,7 +416,7 @@ function _combine_process_pair_symbol(optional_i::Bool,
                                       (firstres,)::Ref{Any},
                                       wfun::Ref{Any},
                                       wincols::Ref{Any},
-                                      multithreaded::Bool)
+                                      threads::Bool)
     @assert only(wfun) isa Base.Callable
     @assert only(wincols) isa Union{Tuple, NamedTuple}
 
@@ -441,7 +441,7 @@ function _combine_process_pair_symbol(optional_i::Bool,
     idx, outcols, _ = _combine_with_first(Ref{Any}(wrap(firstres)), wfun, gd, wincols,
                                           firstmulticol,
                                           firstres isa AbstractVector ? NOTHING_IDX_AGG : idx_agg[],
-                                          multithreaded)
+                                          threads)
     @assert length(outcols) == 1
     outcol = outcols[1]
 
@@ -474,13 +474,13 @@ function _combine_process_pair_astable(optional_i::Bool,
                                        (firstres,)::Ref{Any},
                                        wfun::Ref{Any},
                                        wincols::Ref{Any},
-                                       multithreaded::Bool)
+                                       threads::Bool)
     fun = only(wfun)
     @assert fun isa Base.Callable
     @assert only(wincols) isa Union{Tuple, NamedTuple}
     if firstres isa AbstractVector
         idx, outcol_vec, _ = _combine_with_first(Ref{Any}(wrap(firstres)), wfun, gd, wincols,
-                                                 firstmulticol, NOTHING_IDX_AGG, multithreaded)
+                                                 firstmulticol, NOTHING_IDX_AGG, threads)
         @assert length(outcol_vec) == 1
         res = outcol_vec[1]
         @assert length(res) > 0
@@ -503,7 +503,7 @@ function _combine_process_pair_astable(optional_i::Bool,
             fun = (x...) -> Tables.columntable(oldfun(x...))
         end
         idx, outcols, nms = _combine_multicol(Ref{Any}(firstres), Ref{Any}(fun), gd,
-                                              wincols, multithreaded)
+                                              wincols, threads)
 
         if !(firstres isa Union{AbstractVecOrMat, AbstractDataFrame,
             NamedTuple{<:Any, <:Tuple{Vararg{AbstractVector}}}})
@@ -572,7 +572,7 @@ function _combine_process_pair((cs_i,)::Ref{Any},
                                seen_cols::Dict{Symbol, Tuple{Bool, Int}},
                                trans_res::Vector{TransformationResult},
                                idx_agg::Ref{Vector{Int}},
-                               multithreaded::Bool)
+                               threads::Bool)
     @assert cs_i isa Pair
 
     source_cols, (fun, out_col_name) = cs_i
@@ -596,12 +596,12 @@ function _combine_process_pair((cs_i,)::Ref{Any},
     if out_col_name isa Symbol
         return _combine_process_pair_symbol(optional_i, gd, seen_cols, trans_res, idx_agg,
                                             out_col_name, firstmulticol, Ref{Any}(firstres),
-                                            Ref{Any}(fun), Ref{Any}(incols), multithreaded)
+                                            Ref{Any}(fun), Ref{Any}(incols), threads)
     end
     if out_col_name == AsTable || out_col_name isa AbstractVector{Symbol}
         return _combine_process_pair_astable(optional_i, gd, seen_cols, trans_res, idx_agg,
                                              out_col_name, firstmulticol, Ref{Any}(firstres),
-                                             Ref{Any}(fun), Ref{Any}(incols), multithreaded)
+                                             Ref{Any}(fun), Ref{Any}(incols), threads)
     end
     throw(ArgumentError("unsupported target column name specifier $out_col_name"))
 end
@@ -626,7 +626,7 @@ end
 function _combine(gd::GroupedDataFrame,
                   cs_norm::Vector{Any}, optional_transform::Vector{Bool},
                   copycols::Bool, keeprows::Bool, renamecols::Bool,
-                  multithreaded::Bool)
+                  threads::Bool)
     if isempty(cs_norm)
         if keeprows && nrow(parent(gd)) > 0 && minimum(gd.groups) == 0
             throw(ArgumentError("select and transform do not support " *
@@ -677,7 +677,7 @@ function _combine(gd::GroupedDataFrame,
     for i in eachindex(cs_norm, optional_transform, tasks)
         cs_i = cs_norm[i]
         optional_i = optional_transform[i]
-        tasks[i] = @spawn_or_run_task multithreaded if length(gd) > 0 && isagg(cs_i, gd)
+        tasks[i] = @spawn_or_run_task threads if length(gd) > 0 && isagg(cs_i, gd)
             _combine_process_agg(Ref{Any}(cs_i), optional_i, parentdf, gd,
                                  seen_cols, trans_res, idx_agg[])
         elseif keeprows && cs_i isa Pair && first(last(cs_i)) === identity &&
@@ -688,7 +688,7 @@ function _combine(gd::GroupedDataFrame,
                                   seen_cols, trans_res, idx_keeprows, copycols)
         elseif cs_i isa Base.Callable
             _combine_process_callable(Ref{Any}(cs_i), optional_i, parentdf, gd,
-                                      seen_cols, trans_res, idx_agg, multithreaded)
+                                      seen_cols, trans_res, idx_agg, threads)
         else
             @assert cs_i isa Pair
             if first(cs_i) isa Vector{Int} && isempty(first(cs_i)) &&
@@ -701,7 +701,7 @@ function _combine(gd::GroupedDataFrame,
                                          seen_cols, trans_res, idx_agg[])
             else
                 _combine_process_pair(Ref{Any}(cs_i), optional_i, parentdf, gd,
-                                      seen_cols, trans_res, idx_agg, multithreaded)
+                                      seen_cols, trans_res, idx_agg, threads)
             end
         end
     end
@@ -771,8 +771,8 @@ function _combine(gd::GroupedDataFrame,
 
     @sync for i in eachindex(trans_res)
         let i=i
-            @spawn_or_run multithreaded reorder_cols!(trans_res, i, trans_res[i].col, trans_res[i].col_idx,
-                                                      keeprows, idx_keeprows, gd)
+            @spawn_or_run threads reorder_cols!(trans_res, i, trans_res[i].col, trans_res[i].col_idx,
+                                                keeprows, idx_keeprows, gd)
         end
     end
 
@@ -805,17 +805,17 @@ end
 
 function combine(@nospecialize(f::Base.Callable), gd::GroupedDataFrame;
                  keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-                 multithreaded::Bool=true)
+                 threads::Bool=true)
     if f isa Colon
         throw(ArgumentError("First argument must be a transformation if the second argument is a GroupedDataFrame"))
     end
     return combine(gd, f, keepkeys=keepkeys, ungroup=ungroup, renamecols=renamecols,
-                   multithreaded=multithreaded)
+                   threads=threads)
 end
 
 combine(@nospecialize(f::Pair), gd::GroupedDataFrame;
         keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-        multithreaded::Bool=true) =
+        threads::Bool=true) =
     throw(ArgumentError("First argument must be a transformation if the second argument is a GroupedDataFrame. " *
                         "You can pass a `Pair` as the second argument of the transformation. If you want the return " *
                         "value to be processed as having multiple columns add `=> AsTable` suffix to the pair."))
@@ -824,81 +824,81 @@ combine(gd::GroupedDataFrame,
         @nospecialize(args::Union{Pair, Base.Callable, ColumnIndex, MultiColumnIndex,
                                   AbstractVecOrMat}...);
         keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-        multithreaded::Bool=true) =
+        threads::Bool=true) =
     _combine_prepare(gd, Ref{Any}(map(x -> broadcast_pair(parent(gd), x), args)),
                      keepkeys=keepkeys, ungroup=ungroup,
                      copycols=true, keeprows=false, renamecols=renamecols,
-                     multithreaded=multithreaded)
+                     threads=threads)
 
 function select(@nospecialize(f::Base.Callable), gd::GroupedDataFrame; copycols::Bool=true,
                 keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-                multithreaded::Bool=true)
+                threads::Bool=true)
     if f isa Colon
         throw(ArgumentError("First argument must be a transformation if the second argument is a grouped data frame"))
     end
     return select(gd, f, copycols=copycols, keepkeys=keepkeys, ungroup=ungroup,
-                  multithreaded=multithreaded)
+                  threads=threads)
 end
 
 select(gd::GroupedDataFrame, @nospecialize(args::Union{Pair, Base.Callable, ColumnIndex, MultiColumnIndex,
                                                        AbstractVecOrMat}...);
        copycols::Bool=true, keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-       multithreaded::Bool=true) =
+       threads::Bool=true) =
     _combine_prepare(gd, Ref{Any}(map(x -> broadcast_pair(parent(gd), x), args)),
                      copycols=copycols, keepkeys=keepkeys,
                      ungroup=ungroup, keeprows=true, renamecols=renamecols,
-                     multithreaded=multithreaded)
+                     threads=threads)
 
 function transform(@nospecialize(f::Base.Callable), gd::GroupedDataFrame; copycols::Bool=true,
                    keepkeys::Bool=true, ungroup::Bool=true, renamecols::Bool=true,
-                   multithreaded::Bool=true)
+                   threads::Bool=true)
     if f isa Colon
         throw(ArgumentError("First argument must be a transformation if the second argument is a grouped data frame"))
     end
     return transform(gd, f, copycols=copycols, keepkeys=keepkeys, ungroup=ungroup,
-                     multithreaded=multithreaded)
+                     threads=threads)
 end
 
 function transform(gd::GroupedDataFrame, @nospecialize(args::Union{Pair, Base.Callable, ColumnIndex, MultiColumnIndex,
                                                                    AbstractVecOrMat}...);
                    copycols::Bool=true, keepkeys::Bool=true, ungroup::Bool=true,
                    renamecols::Bool=true,
-                   multithreaded::Bool=true)
+                   threads::Bool=true)
     res = select(gd, :, args..., copycols=copycols, keepkeys=keepkeys,
-                 ungroup=ungroup, renamecols=renamecols, multithreaded=multithreaded)
+                 ungroup=ungroup, renamecols=renamecols, threads=threads)
     # res can be a GroupedDataFrame based on DataFrame or a DataFrame,
     # so parent always gives a data frame
-    select!(parent(res), propertynames(parent(gd)), :, multithreaded=multithreaded)
+    select!(parent(res), propertynames(parent(gd)), :, threads=threads)
     return res
 end
 
 function select!(@nospecialize(f::Base.Callable), gd::GroupedDataFrame;
-                 ungroup::Bool=true, renamecols::Bool=true, multithreaded::Bool=true)
+                 ungroup::Bool=true, renamecols::Bool=true, threads::Bool=true)
     if f isa Colon
         throw(ArgumentError("First argument must be a transformation if the second argument is a grouped data frame"))
     end
-    return select!(gd, f, ungroup=ungroup, multithreaded=multithreaded)
+    return select!(gd, f, ungroup=ungroup, threads=threads)
 end
 
 function select!(gd::GroupedDataFrame,
                  @nospecialize(args::Union{Pair, Base.Callable, ColumnIndex, MultiColumnIndex,
                                            AbstractVecOrMat}...);
-                 ungroup::Bool=true, renamecols::Bool=true, multithreaded::Bool=true)
+                 ungroup::Bool=true, renamecols::Bool=true, threads::Bool=true)
     df = parent(gd)
     if df isa DataFrame
         newdf = select(gd, args..., copycols=false, renamecols=renamecols,
-                       multithreaded=multithreaded)
+                       threads=threads)
     else
         @assert df isa SubDataFrame
         newdf = select(gd, args..., copycols=true, renamecols=renamecols,
-                       multithreaded=multithreaded)
+                       threads=threads)
     end
     _replace_columns!(df, newdf)
     return ungroup ? df : gd
 end
 
 function transform!(@nospecialize(f::Base.Callable), gd::GroupedDataFrame;
-                    ungroup::Bool=true, renamecols::Bool=true, multithreaded::Bool=true)
+                    ungroup::Bool=true, renamecols::Bool=true, threads::Bool=true)
     if f isa Colon
         throw(ArgumentError("First argument must be a transformation if the second argument is a grouped data frame"))
     end
@@ -908,17 +908,17 @@ end
 function transform!(gd::GroupedDataFrame,
                     @nospecialize(args::Union{Pair, Base.Callable, ColumnIndex, MultiColumnIndex,
                                               AbstractVecOrMat}...);
-                    ungroup::Bool=true, renamecols::Bool=true, multithreaded::Bool=true)
+                    ungroup::Bool=true, renamecols::Bool=true, threads::Bool=true)
     df = parent(gd)
     if df isa DataFrame
         newdf = select(gd, :, args..., copycols=false, renamecols=renamecols,
-                       multithreaded=multithreaded)
+                       threads=threads)
     else
         @assert df isa SubDataFrame
         newdf = select(gd, :, args..., copycols=true, renamecols=renamecols,
-                       multithreaded=multithreaded)
+                       threads=threads)
     end
-    select!(newdf, propertynames(df), :, multithreaded=multithreaded)
+    select!(newdf, propertynames(df), :, threads=threads)
     _replace_columns!(df, newdf)
     return ungroup ? df : gd
 end